create Folder const and file barrel
