+++ src/components/DonutChart.jsx
@@ -1,42 +1,11 @@
-'use client'
+"use client";
 
-import Chart from "react-apexcharts"
+import { OPTIONS } from "@/app/const/varDonutChart";
+import Chart from "react-apexcharts";
 
-const OPTIONS = {
-  legend: {
-    show: false
-  },
-  chart: {
-    height: 350,
-    type: 'donut'
-  },
-  tooltip: {
-    enabled: true,
-    formatter: function (val) {
-      return val
-    },
-    style: {
-      color: '#181818',
-      fontSize: '12px',
-      background: '#181818',
-    },
-    theme: false,
-    x: {
-      show: true,
-      formatter: undefined,
-    },
-    y: {
-      formatter: undefined,
-      title: {
-        formatter: (seriesName) => seriesName,
-      },
-    },
-  }
-}
-
-export default function DonutChart ({ data }) {
+export default function DonutChart({ data }) {
   return (
-    <section className='w-[80%] mx-auto'>
+    <section className="w-[80%] mx-auto">
       <Chart
         type="donut"
         width={380}
@@ -44,5 +13,5 @@ export default function DonutChart ({ data }) {
         series={data.series}
       />
     </section>
-  )
+  );
 }
