+++ src/components/index.js
@@ -1,0 +1,5 @@
+export { default as DonutChart } from './DonutChart';
+export { default as ReviewCard } from './ReviewCard';
+export { default as ThemeSwitcher } from './ThemeSwitcher';
+export { default as Timer } from './Timer';
+export { default as TreeMapChart } from './TreeMapChart';
